+++ app/(route)/join/page.tsx
@@ -4,8 +4,7 @@ import Button from '@/components/common/Button';
 import Input from '@/components/common/Input';
 import Timer from '@/components/common/Timer';
 import CustomToast from '@/components/common/Toast';
-import useCheckAuthCode from '@/hooks/queries/useCheckAuthcode';
-import useMailNumber from '@/hooks/queries/useGetAuthcode';
+import { useCheckAuthCode, useMailNumber } from '@/hooks/queries/useAuthCode';
 import { useJoin } from '@/hooks/queries/useAuth';
 import { useRouter } from 'next/navigation';
 import React, { useEffect, useState } from 'react';
